@@ -7,6 +7,7 @@
   "homepage": "https://mersocarlin.com",
   "private": true,
   "sideEffects": false,
+  "type": "module",
   "scripts": {
     "build": "run-s \"build:*\"",
     "build:css": "npm run generate:css -- --minify",
@@ -15,17 +16,21 @@
     "dev:css": "npm run generate:css -- --watch",
     "dev:remix": "remix dev",
     "generate:css": "npx tailwindcss -i ./styles/tailwind.css -o ./app/tailwind.css",
-    "typecheck": "tsc -b"
+    "typecheck": "tsc -b",
+    "start": "cross-env NODE_ENV=production remix-serve ./build/index.js"
   },
   "dependencies": {
     "@amplitude/node": "^1.10.2",
-    "@remix-run/node": "^1.10.0",
-    "@remix-run/react": "^1.10.0",
-    "@remix-run/vercel": "^1.10.0",
+    "@remix-run/css-bundle": "^2.2.0",
+    "@remix-run/node": "^2.2.0",
+    "@remix-run/react": "^2.2.0",
+    "@remix-run/serve": "^2.8.1",
     "@vercel/node": "^2.6.2",
     "clsx": "^1.2.1",
+    "cross-env": "^7.0.3",
     "date-fns": "^2.29.3",
-    "mdx-bundler": "^9.2.1",
+    "esbuild": "^0.20.2",
+    "mdx-bundler": "^10.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "reading-time": "^1.5.0",
@@ -35,9 +40,8 @@
     "tiny-invariant": "^1.3.1"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.10.0",
-    "@remix-run/eslint-config": "^1.10.0",
-    "@remix-run/serve": "^1.10.0",
+    "@remix-run/dev": "^2.2.0",
+    "@remix-run/eslint-config": "^2.2.0",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/line-clamp": "^0.4.2",
     "@tailwindcss/typography": "^0.5.9",
@@ -47,7 +51,7 @@
     "eslint": "^8.27.0",
     "npm-run-all": "^4.1.5",
     "tailwindcss": "^3.2.4",
-    "typescript": "^4.8.4"
+    "typescript": "^5.2.2"
   },
   "engines": {
     "node": ">=16"
